Api ordenes por ID

+++ src/main/java/com/unir/movie_app_operator/persistence/entity/OrdenesEntity.java
@@ -13,7 +13,7 @@ public class OrdenesEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long ordenID;
+    private Integer ordenID;
 
     @ManyToOne
     @JoinColumn(name = "id", nullable = false)
